requirements.txt
Django==5.2.4
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
faiss-cpu==1.7.4
PyPDF2==3.0.1
python-dotenv==1.0.0
openai==1.6.1
tiktoken==0.5.2
numpy==1.24.3
scikit-learn==1.3.0